feat(template): add new forward-call template
