+++ forward-call/.env
@@ -1,0 +1,2 @@
+# The number you want to forward incoming calls to
+MY_PHONE_NUMBER=+12223334444
