--- never-gonna-give-you-up/README.md
@@ -1,13 +1,0 @@
-# Never gonna give you up
-
-This Function will never let you down to play Rick Astley's hit song! 
-
-You can use this template to play your own song as well.
-
-## Environment variables
-
-This Function requires no environment variables to run successfully.
-
-## Parameters
-
-This Function requires no URL or POST parameters to run successfully.
