--- never-gonna-give-you-up/never-gonna-give-you-up.js
@@ -1,6 +1,0 @@
-exports.handler = function(context, event, callback) {
-  const twiml = new Twilio.twiml.VoiceResponse();
-  twiml.play('https://demo.twilio.com/docs/classic.mp3');
-
-  callback(null, twiml);
-};
