+++ templates.json
@@ -11,9 +11,9 @@
       "description": "Basic Function that will return Hello World"
     },
     {
-      "id": "never-gonna-give-you-up",
-      "name": "Never gonna give you up",
-      "description": "Never gonna let you down. Plays a song to a phone call"
+      "id": "forward-call",
+      "name": "Forward Call",
+      "description": "Forwards an incoming call to another number"
     },
     {
       "id": "forward-message",
